Refactor - create separate spec/folder for each resource type - makes navigating specs easier
