+++ spec/RDS/RDS.rds.canMakeConnection.spec.ts
@@ -1,6 +1,6 @@
-import { RDS } from '../src/RDS'
+import { RDS } from '../../src/RDS'
 import { connectionTester, endpointAddress } from '../src/interfaces'
-import { callbackSuccessReturning, nvp } from './support'
+import { callbackSuccessReturning, nvp } from '../support'
 import { testRds } from './RDS.stub'
 var AWSMock = require('aws-sdk-mock')
 var sinon = require('sinon')
