+++ spec/S3/S3.bucket.hasAccessLogging.spec.ts
@@ -1,5 +1,5 @@
-import { S3 } from "../src/S3"
-import { awsMockCallback, awsMockFailureCallback } from './support'
+import { S3 } from '../../src/S3'
+import { awsMockCallback, awsMockFailureCallback } from '../support'
 import AWSMock = require('aws-sdk-mock')
 
 describe('S3.bucket#hasAccessLogging', () => {
