+++ spec/VPC/VPC.stub.ts
